feat(sc-sets): add abstract sc set
